+++ package.json
@@ -16,6 +16,7 @@
     "@trpc/next": "^10.8.1",
     "@trpc/react-query": "^10.8.1",
     "@trpc/server": "^10.8.1",
+    "clsx": "^1.2.1",
     "next": "13.1.1",
     "next-auth": "^4.18.8",
     "react": "18.2.0",
@@ -24,6 +25,7 @@
     "zod": "^3.20.2"
   },
   "devDependencies": {
+    "@svgr/webpack": "^6.5.1",
     "@types/node": "^18.11.18",
     "@types/prettier": "^2.7.2",
     "@types/react": "^18.0.26",
